Ejercicio 14.1

+++ Ejercicio 14/Ejercicio 14/Program.cs
@@ -1,12 +1,12 @@
-﻿int numero = -1;
-int numeromayor = int.MinValue;
+﻿double numero = -1;
+double numeromayor = double.MinValue;
 Console.WriteLine("Ingrese los números que quiera e ingrese 0 para mostrar la suma final");
 Console.WriteLine();
 do
 {
     Console.WriteLine($"Ingrese un número");
     Console.WriteLine();
-    numero = int.Parse(Console.ReadLine());
+    numero = double.Parse(Console.ReadLine());
     Console.WriteLine();
     if (numeromayor<numero)
     {
